allowing for individual icon export, flagging current icon component as deprecated for external consumption
