+++ min-coverage.json
@@ -1,1 +1,1 @@
-{"total":1146,"covered":1094,"skipped":0,"pct":95.46}
+{"total":1247,"covered":1195,"skipped":0,"pct":95.83}
