+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "he-react-ui",
-  "version": "3.0.6",
+  "version": "3.0.7",
   "description": "Common UI elements for HealthEngine",
   "main": "dist/index.js",
   "scripts": {
