+++ src/components/Form/Button/fixtures/iconRight.fixture.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
 import Button from '../';
-import Icon from '../../../Icon';
+import { IconArrowRight } from '../../../Icon';
 
 export default {
   component: Button,
   props: {
-    iconRight: <Icon name="ArrowRight" />,
+    iconRight: <IconArrowRight />,
   },
   children: 'Button',
   simulateSubmission: true,
