+++ src/components/index.tsx
@@ -1,4 +1,4 @@
-import Icon from './Icon';
+import Icon from './Icon/Icon';
 import Notification from './Notification';
 import Popover from './Popover';
 import PopUp from './PopUp';
@@ -6,4 +6,5 @@ import Table from './Table';
 
 export * from './Form';
 export * from './Loading';
+export * from './Icon';
 export { Icon, Notification, Popover, PopUp, Table };
